+++ R/random_mpm.R
@@ -59,7 +59,7 @@
 #'   determined by the archetype selected and fecundity vector. If split = TRUE,
 #'   the matrix is split into separate fertility and a growth/survival matrices,
 #'   returned as a list.
-#' @family simulation
+#' @family Lefkovitch matrices
 #' @author Owen Jones <jones@biology.sdu.dk>
 #'
 #' @importFrom MCMCpack rdirichlet