+++ NEWS.md
@@ -1,3 +1,8 @@
+# mpmsim (development version)
+
+- added function to simulate the action of drivers such as weather on vital rates: `drive_vital_rate()`.
+- added full suite of unit tests using `testthat`. Test coverage 97.01%.
+
 # mpmsim 1.0.0
 
 ### First release, with the following functionality.